requirements.txt
discord.py==1.3.4
redis==3.5.3